Add basis for player paddle drawing
To begin the process of developing an in-game player input controller, this
prototype uses the LineRenderer component to draw lines spanning from the users
mouse-down coordinate to the coordinate of the mouse-up event.

The next step will involve making these rendered lines act as colliders; and if
successful, the PlayerInput controller will be extended to add features such as
max line length and a limit on the number of drawable lines.